\name{argostools-package}
\alias{argostools-package}
\alias{argostools}
\docType{package}
\title{
Creating and managing an SQL framework for Argos locations data
}
\description{
This package provides functions to import (xls(x), txt, csv and diag extensions) and format Argos data (coordinates and date fields). Some filters are also available (spatial, temporal and speed-based filters) to clean up locations. After compilation, data can be passed under an SQL structure using several predefined functions. Several other tools are also available.
}
\details{
\tabular{ll}{
Package: \tab argostools\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2015-04-14\cr
License: \tab GPL-2\cr
}
}
\author{
Nicolas Casajus

Maintainer: Nicolas Casajus <nicolas.casajus@gmail.com>
}
\keyword{ package }
\seealso{
\code{\link[XLConnect:XLConnect-package]{XLConnect}}, \code{\link[RPostgreSQL:PostgreSQL]{RPostgreSQL}}, \code{\link[sp:sp]{sp}}
}
