% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_add_new_ptt.R
\name{pg_add_new_ptt}
\alias{pg_add_new_ptt}
\title{Get Animal PTT information from PostgreSQL database}
\usage{
pg_add_new_ptt(dbname, noid, ptt, start)
}
\arguments{
\item{dbname}{A DBI connexion information}

\item{noid}{ID of the animal}

\item{ptt}{ID of the PTT}

\item{start}{Starting date of tracking of the PTT}
}
\description{
Get Animal PTT information from PostgreSQL database.
}
\examples{
## See vignette

}
