% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_get_id_by_ptt.R
\name{pg_get_id_by_ptt}
\alias{pg_get_id_by_ptt}
\title{Get Animal ID from PTT}
\usage{
pg_get_id_by_ptt(dbname, ptt)
}
\arguments{
\item{dbname}{A DBI connexion information}

\item{ptt}{ID of one or several PTT(s)}
}
\value{
A data frame with 2 variables
}
\description{
Get Animal ID from PTT stored in a PostgreSQL database.
}
\examples{
## See vignette

}
