% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_get_ptts.R
\name{pg_get_ptts}
\alias{pg_get_ptts}
\title{Get Animal PTT information from PostgreSQL database}
\usage{
pg_get_ptts(dbname, noid = NULL)
}
\arguments{
\item{dbname}{A DBI connexion information}

\item{noid}{ID of the animal. If missing, all ID are processed}
}
\value{
A data frame with 4 variables
}
\description{
Get Animal PTT information from PostgreSQL database.
}
\examples{
## See vignette

}
