% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_rm_duplicates.R
\name{pg_rm_duplicates}
\alias{pg_rm_duplicates}
\title{Remove duplicates in PostgreSQL relation}
\usage{
pg_rm_duplicates(dbname, tbname)
}
\arguments{
\item{dbname}{A DBI connexion information}

\item{tbname}{PostgreSQL table name in which duplicates will be removed}
}
\description{
Remove duplicates in PostgreSQL relation.
}
\examples{
## See vignette

}
