% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_update_tb.R
\name{pg_update_tb}
\alias{pg_update_tb}
\title{Update/create a PostgreSQL table}
\usage{
pg_update_tb(data = NULL, dbname, tbname, fields)
}
\arguments{
\item{data}{A data frame with data to insert}

\item{dbname}{A DBI connexion information}

\item{tbname}{PostgreSQL table name}

\item{fields}{A list  of  fields}
}
\description{
Update/create a PostgreSQL table.
}
\examples{
## See vignette

}
