\name{pgextractlocs}
\alias{pgextractlocs}
\title{SQL query to extract Argos locations}

\description{Description}

\usage{
pgextractlocs(dbname, id, byid = T, start = NULL, end = NULL)
}

\arguments{
  \item{dbname}{}
  \item{id}{}
  \item{byid}{}
  \item{start}{}
  \item{end}{}
}

\value{}

\author{Nicolas Casajsu}

\seealso{}

\examples{
###
### See tutorial for a complete example of usage of this package.
}
