\name{pggetidbyptt}
\alias{pggetidbyptt}
\title{
}
\description{
}
\usage{
pggetidbyptt(dbname, ptt)
}
\arguments{
  \item{dbname}{
}
  \item{ptt}{
}
}
\details{
}
\value{
}
\references{
}
\author{
}
\note{
}

\seealso{
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
}
