\name{pgrmduplicates}
\alias{pgrmduplicates}
\title{Title}
\description{Description}

\usage{
pgrmduplicates(dbname, tbname)
}

\arguments{
  \item{dbname}{}
  \item{tbname}{}
}

\value{}

\author{Nicolas Casajus}

\seealso{}

\examples{
###
### See tutorial for a complete example of usage of this package.
}
