% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_argos.R
\name{read_argos}
\alias{read_argos}
\title{Import and clean raw Argos locations data}
\usage{
read_argos(file)
}
\arguments{
\item{file}{Path to the filename to read}
}
\value{
A data frame with 7 variables
}
\description{
Import and clean raw Argos locations data.
}
\examples{
## See vignette

}
