% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_buffer.R
\name{spatial_buffer}
\alias{spatial_buffer}
\title{Remove Argos location outside a spatial buffer}
\usage{
spatial_buffer(data, xlim, ylim, lat = "latitude", long = "longitude")
}
\arguments{
\item{data}{Cleaned Argos location data frame}

\item{xlim}{Min and Max of longitude of the spatial buffer}

\item{ylim}{Min and Max of latitude of the spatial buffer}

\item{lat}{Column name of the latitude field}

\item{long}{Column name of the longitude field}
}
\value{
A data frame with 7 variables
}
\description{
Remove Argos location outside a spatial buffer set by user.
}
\examples{
## See vignette

}
