% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed_filter.R
\name{speed_filter}
\alias{speed_filter}
\title{Remove Argos location based on speed characteristics}
\usage{
speed_filter(
  data,
  selqual = c(0, 1, 2, 3),
  speed1 = 7,
  speed2 = 10,
  timespeed2 = 12,
  noid = "noid",
  date = "dateloc",
  long = "longitude",
  lat = "latitude",
  qual = "quality",
  crs = "+proj=utm +zone=17 +datum=NAD83 +ellps.default=GCS"
)
}
\arguments{
\item{data}{Cleaned Argos location data frame}

\item{selqual}{Location qualities to keep}

\item{speed1}{Lower speed}

\item{speed2}{Upper speed}

\item{timespeed2}{Duration of upper speed}

\item{noid}{Column name of animal id}

\item{date}{Column name of location date}

\item{long}{Column name of longitude}

\item{lat}{Column name of latitude}

\item{qual}{Column name of location quality}

\item{crs}{proj4string of UTM zone of study area}
}
\value{
A data frame with 12 variables
}
\description{
Remove Argos location based on speed characteristics.
}
\examples{
## See vignette

}
