\name{speedfilter}
\alias{speedfilter}
\title{
}
\description{
}
\usage{
speedfilter(data, selqual = c(0, 1, 2, 3), speed1 = 7, speed2 = 10, timespeed2 = 12, noid = "noid", date = "dateloc", long = "longitude", lat = "latitude", qual = "quality", crs = "+proj=utm +zone=17 +datum=NAD83 +ellps.default=GCS")
}
\arguments{
  \item{data}{
}
  \item{selqual}{
}
  \item{speed1}{
}
  \item{speed2}{
}
  \item{timespeed2}{
}
  \item{noid}{
}
  \item{date}{
}
  \item{long}{
}
  \item{lat}{
}
  \item{qual}{
}
  \item{crs}{
}
}
\details{
}
\value{
}
\references{
}
\author{
}
\note{
}

\seealso{
}
\examples{
}