% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_files.R
\name{stack_files}
\alias{stack_files}
\title{Import and clean raw Argos locations data (multiple files)}
\usage{
stack_files(files)
}
\arguments{
\item{files}{Vector of filenames to read}
}
\value{
A data frame with 7 variables
}
\description{
Import and clean raw Argos locations data(multiple files).
}
\examples{
## See vignette

}
