% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_buffer.R
\name{temporal_buffer}
\alias{temporal_buffer}
\title{Remove Argos location inside a temporal buffer}
\usage{
temporal_buffer(
  data,
  interval = 5,
  noid = "noid",
  date = "dateloc",
  qual = "quality",
  semmajaxis = "semmajaxis"
)
}
\arguments{
\item{data}{Cleaned Argos location data frame}

\item{interval}{Length of the temporal buffer (in minutes) inside which locations are considered to be redundondant}

\item{noid}{Column name of the PTT field}

\item{date}{Column name of the date field}

\item{qual}{Column name of the location quality field}

\item{semmajaxis}{Column name of the semi-major axis field}
}
\value{
A data frame with 7 variables
}
\description{
Remove Argos location inside a temporal buffer set by user.
}
\examples{
## See vignette

}
