\name{temporalbuffer}
\alias{temporalbuffer}
\title{Title}
\description{Description}

\usage{
temporalbuffer(data, interval = 5, noid = "noid", date = "dateloc", qual = "quality", semmajaxis = "semmajaxis")
}

\arguments{
  \item{data}{}
  \item{interval}{}
  \item{noid}{}
  \item{date}{}
  \item{qual}{}
  \item{semmajaxis}{}
}

\value{}

\author{Nicolas Casajus}

\seealso{}

\examples{
###
### See tutorial for a complete example of usage of this package.
}
