% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_length.R
\name{track_length}
\alias{track_length}
\title{Get duration of animal tracking}
\usage{
track_length(data, id, noid = "noid", date = "dateloc", unit = "days")
}
\arguments{
\item{data}{Cleaned Argos location data frame}

\item{id}{The animal id to extract track informations}

\item{noid}{Column name of the animal id field}

\item{date}{Column name of the date field}

\item{unit}{Default is 'days'}
}
\value{
A data frame with 7 variables
}
\description{
Get duration of animal tracking.
}
\examples{
## See vignette

}
