% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argostools-package.R
\docType{package}
\name{argostools-package}
\alias{argostools}
\alias{argostools-package}
\title{argostools: Creating and managing SQL framework for Argos locations data}
\description{
This package provides functions to import (xls(x), txt, csv and diag extensions) and format Argos data (coordinates and date fields). Some filters are also available (spatial, temporal and speed-based filters) to clean up locations. After compilation, data can be passed under an SQL structure using several functions (database creation, relation update, extraction queries, etc.)
}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@gmail.com}

}
\keyword{internal}
