% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_extract_locs.R
\name{pg_extract_locs}
\alias{pg_extract_locs}
\title{Get Argos locations}
\usage{
pg_extract_locs(dbname, id = NULL, start = NULL, end = NULL)
}
\arguments{
\item{dbname}{A DBI connexion information}

\item{id}{Animal ID (not PTT)}

\item{start}{Minimum location date}

\item{end}{Maximum location date}
}
\value{
A data frame with 7 variables
}
\description{
Get Argos locations for one or several animals from a PostgreSQL database.
}
\examples{
## See vignette

}
