\name{readargos}
\alias{readargos}
\title{Title}
\description{Description}

\usage{
readargos(file, sep = "\t", dec = ".", header = TRUE, ...)
}

\arguments{
  \item{file}{}
  \item{sep}{}
  \item{dec}{}
  \item{header}{}
  \item{\dots}{}
}

\value{}

\author{Nicolas Casajus}

\seealso{}

\examples{
###
### See tutorial for a complete example of usage of this package.
}
