\name{temporalbuffer}
\alias{temporalbuffer}
\title{
}
\description{
}
\usage{
temporalbuffer(data, interval = 5, noid = "noid", date = "dateloc", qual = "quality", semmajaxis = "semmajaxis")
}
\arguments{
  \item{data}{
}
  \item{interval}{
}
  \item{noid}{
}
  \item{date}{
}
  \item{qual}{
}
  \item{semmajaxis}{
}
}
\details{
}
\value{
}
\references{
}
\author{
}
\note{
}

\seealso{
}
\examples{
}