\name{tracklen}
\alias{tracklen}
\title{Title}
\description{Description}

\usage{
tracklen(data, id, noid = "noid", date = "dateloc", unit = "days")
}

\arguments{
  \item{data}{}
  \item{id}{}
  \item{noid}{}
  \item{date}{}
  \item{unit}{}
}

\value{}

\author{Nicolas Casajsu}

\seealso{}

\examples{
###
### See tutorial for a complete example of usage of this package.
}
